Prevent int division for multiplier calculation

+++ src/main/java/com/c2t2s/hb/AllOrNothing.java
@@ -201,7 +201,7 @@ class AllOrNothing {
         }
 
         double getPayoutMultiplier() {
-            return Math.pow(2, rolls / difficulty.rollsToDouble);
+            return Math.pow(2, ((double)rolls) / difficulty.rollsToDouble);
         }
 
         boolean isClaimable() {
